fixed 'Route' url, now will open the cards
